Migrating changes to Wallet
