+++ packages/cosmos-bin/src/cli.rs
@@ -119,9 +119,4 @@ pub(crate) enum Subcommand {
         #[clap(subcommand)]
         opt: crate::config::Opt,
     },
-    /// Management of cryptographic operations
-    Crypto {
-        #[clap(flatten)]
-        opt: crate::crypto::Opt,
-    },
 }
