+++ packages/cosmos/src/lib.rs
@@ -10,7 +10,6 @@ pub use cosmos_builder::CosmosBuilder;
 pub use cosmos_network::CosmosNetwork;
 pub use cosmos_sdk_proto as proto;
 pub use cosmos_sdk_proto::cosmos::base::v1beta1::Coin;
-pub use crypto::K256;
 pub use error::Error;
 pub use ext::TxResponseExt;
 pub use gas_multiplier::DynamicGasMultiplier;
@@ -28,7 +27,6 @@ mod config;
 mod contract;
 mod cosmos_builder;
 mod cosmos_network;
-mod crypto;
 mod ext;
 mod gas_multiplier;
 mod injective;
